+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.3.7...HEAD) - TBD
 
+### :wrench: Changed
+
+- Display mutation metadata in `negative_data_rejection` failures if available. Currently, only available for the coverage phase.
+
 ## [4.3.7](https://github.com/schemathesis/schemathesis/compare/v4.3.6...v4.3.7) - 2025-10-20
 
 ### :bug: Fixed